@@ -62,17 +62,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Shop Model
-    |--------------------------------------------------------------------------
-    |
-    | This option is for overriding the shop model with your own.
-    |
-    */
-
-    'shop_model' => env('SHOPIFY_SHOP_MODEL', '\OhMyBrew\ShopifyApp\Models\Shop'),
-
-    /*
-    |--------------------------------------------------------------------------
     | AppBridge Mode
     |--------------------------------------------------------------------------
     |
@@ -109,7 +98,7 @@ return [
     |
     */
 
-    'api_version' => env('SHOPIFY_API_VERSION', null),
+    'api_version' => env('SHOPIFY_API_VERSION', '2020-01'),
 
     /*
     |--------------------------------------------------------------------------
