@@ -1,19 +1,22 @@
 <?php
 
 use Faker\Generator as Faker;
-use OhMyBrew\ShopifyApp\Models\Shop;
+use Illuminate\Support\Facades\Config;
 
-$factory->define(Shop::class, function (Faker $faker) {
+$model = Config::get('auth.providers.users.model');
+
+$factory->define($model, function (Faker $faker) {
     return [
-        'shopify_domain' => "{$faker->domainWord}.myshopify.com",
-        'shopify_token'  => str_replace('-', '', $faker->uuid),
+        'name'     => "{$faker->domainWord}.myshopify.com",
+        'password' => str_replace('-', '', $faker->uuid),
+        'email'    => '',
     ];
 });
 
-$factory->state(Shop::class, 'freemium', [
-    'freemium' => true,
+$factory->state($model, 'freemium', [
+    'shopify_freemium' => true,
 ]);
 
-$factory->state(Shop::class, 'grandfathered', [
-    'grandfathered' => true,
+$factory->state($model, 'grandfathered', [
+    'shopify_grandfathered' => true,
 ]);
